Enum av::LogLevel

#[repr(i32)]

pub enum LogLevel {
    Quiet,
    Panic,
    Fatal,
    Error,
    Warning,
    Info,
    Verbose,
    Debug,
    Trace,
}

Variants
Quiet

Print no output.

Panic

Something went really wrong and we will crash now.

Fatal

Something went wrong and recovery is not possible. For example, no header was found for a format which depends on headers or an illegal combination of parameters is used.

Error

Something went wrong and cannot losslessly be recovered. However, not all future data is affected.

Warning

Something somehow does not look correct. This may or may not lead to problems. An example would be the use of '-vstrict -2'.

Info

Standard information.

Verbose

Detailed information.

Debug

Stuff which is only useful for libav* developers.

Trace

Extremely verbose debugging, useful for libav* development.