Struct av::video::Scaler

pub struct Scaler { /* fields omitted */ }

A context for scaling/converting video frames.

Scaling is most efficiently done by reusing the scaler for a specific combination of source/target width, height and format. It's allowed to use different values for each invocation of the scaling functions, but it will result in reallocation of the internal scaling context, which might not be desireable.

Methods

impl Scaler

fn new() -> Self

Create a new scaling context.

fn scale(
&mut self,
src_data: &[&[u8]],
src_linesize: &[usize],
src_w: usize,
src_h: usize,
src_fmt: AVPixelFormat,
dst_data: &mut [&mut [u8]],
dst_linesize: &[usize],
dst_w: usize,
dst_h: usize,
dst_fmt: AVPixelFormat
) -> Result<()>

Scale src_data to src_data by using the given dimensions and formats.

Requirements

src_fmt and dst_fmt need to be valid pixel formats.

src_w, src_h, dst_w and dst_h need to be greater than 0.

The number of planes need to be greater than or equal to the number of planes required by the pixel formats.

The planes need to be big enough to contain the amount of bytes described by their linesize and the height (linesize * height).

fn scale_frame(&mut self, src: &Frame, dst: &mut Frame) -> Result<()>

Copy the src pixel data to the dst pixel data, scaling dimensions and converting pixel formats as required.